# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EuLLM Agent Core API",
    "version": "0.3.0",
    "description": "Runs, single model calls, approvals and checked HTTP fetches. Every endpoint except health and this document needs `Authorization: Bearer <token>`. Data is scoped to the tenant of the token."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8088"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CreateRun": {
        "type": "object",
        "required": [
          "input"
        ],
        "properties": {
          "profile": {
            "type": "string",
            "default": "default",
            "description": "Run profile from the configuration"
          },
          "input": {
            "type": "string",
            "maxLength": 102400,
            "description": "The task"
          }
        }
      },
      "RunAccepted": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued"
            ]
          }
        }
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "waiting_approval",
          "succeeded",
          "failed"
        ]
      },
      "RunSummary": {
        "type": "object",
        "required": [
          "id",
          "tenant",
          "profile",
          "source",
          "status",
          "created_at_ms",
          "finished_at_ms"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "created_at_ms": {
            "type": "integer"
          },
          "finished_at_ms": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "RunList": {
        "type": "object",
        "required": [
          "runs"
        ],
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            }
          }
        }
      },
      "Step": {
        "type": "object",
        "required": [
          "kind",
          "at_ms"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "llm_call",
              "tool_call"
            ]
          },
          "at_ms": {
            "type": "integer"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "duration_ms": {
            "type": "integer"
          },
          "input_tokens": {
            "type": [
              "integer",
              "null"
            ]
          },
          "output_tokens": {
            "type": [
              "integer",
              "null"
            ]
          },
          "cost": {
            "type": [
              "number",
              "null"
            ]
          },
          "tool": {
            "type": "string"
          },
          "argument_keys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "decision": {
            "type": "string",
            "enum": [
              "allow",
              "deny",
              "require_approval"
            ]
          },
          "decision_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "approval_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "output_bytes": {
            "type": [
              "integer",
              "null"
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "Run": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RunSummary"
          },
          {
            "type": "object",
            "required": [
              "input",
              "output",
              "error",
              "iterations",
              "input_tokens",
              "output_tokens",
              "cost",
              "tainted",
              "steps"
            ],
            "properties": {
              "input": {
                "type": "string"
              },
              "output": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "error": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "iterations": {
                "type": "integer"
              },
              "input_tokens": {
                "type": "integer"
              },
              "output_tokens": {
                "type": "integer"
              },
              "cost": {
                "type": [
                  "number",
                  "null"
                ]
              },
              "tainted": {
                "type": "boolean",
                "description": "The run read external content"
              },
              "steps": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Step"
                }
              }
            }
          }
        ]
      },
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "messages"
        ],
        "properties": {
          "model": {
            "type": "string",
            "default": "default",
            "description": "Model name from the Model Router"
          },
          "messages": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatTool"
            },
            "description": "Tool definitions the model may call; the Core does not execute them here"
          }
        }
      },
...
      "ChatResponse": {
        "type": "object",
        "required": [
          "content",
          "tool_calls",
          "usage",
          "cost"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "tool_calls": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "name",
                "arguments"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "arguments": {}
              }
            }
          },
          "usage": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          },
          "cost": {
            "type": [
              "number",
              "null"
            ]
          }
        }
      },
      "Approval": {
        "type": "object",
        "required": [
          "id",
          "tenant",
          "run_id",
          "tool",
          "arguments",
          "reason",
          "status",
          "decided_by",
          "decision_note",
          "created_at_ms",
          "decided_at_ms"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant": {
            "type": "string"
          },
          "run_id": {
            "type": "string",
            "format": "uuid"
          },
          "tool": {
            "type": "string"
          },
          "arguments": {
            "type": "object",
            "description": "Values shown to the person deciding (long strings shortened)"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "denied",
              "expired"
            ]
          },
          "decided_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "decision_note": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at_ms": {
            "type": "integer"
          },
          "decided_at_ms": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      },
      "ApprovalList": {
        "type": "object",
        "required": [
          "approvals"
        ],
        "properties": {
          "approvals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approval"
            }
          }
        }
      },
      "Decision": {
        "type": "object",
        "required": [
          "decision"
        ],
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approve",
              "deny"
            ]
          },
          "note": {
            "type": "string"
          }
        }
      },
      "FetchRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "https URL (http only when allowed)"
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Only accept, accept-language, if-none-match, if-modified-since"
          }
        }
      },
      "FetchResponse": {
        "type": "object",
        "required": [
          "url",
          "status",
          "body_base64",
          "truncated",
          "redirects"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "Final URL after redirects"
          },
          "status": {
            "type": "integer"
          },
          "content_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "etag": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_modified": {
            "type": [
              "string",
              "null"
            ]
          },
          "body_base64": {
            "type": "string"
          },
          "truncated": {
            "type": "boolean",
            "description": "The body stopped at the size limit"
          },
          "redirects": {
            "type": "integer"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/v1/runs": {
      "post": {
        "summary": "Start a run",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRun"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Run accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or unknown profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Profile not allowed for this token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List recent runs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Runs, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{id}": {
      "get": {
        "summary": "A run with its steps",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No such run for this tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/llm/chat": {
      "post": {
        "summary": "One model call through the Model Router, recorded with usage and cost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or unknown model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Model not allowed for this token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The provider failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals": {
      "get": {
        "summary": "Approvals of this tenant",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "denied",
                "expired",
                "all"
              ],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approvals, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalList"
                }
              }
            }
          },
          "400": {
            "description": "Unknown status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals/{id}": {
      "post": {
        "summary": "Approve or deny a pending action",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The decided approval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approval"
                }
              }
            }
          },
          "400": {
            "description": "Invalid decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Not found or already decided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fetch": {
      "post": {
        "summary": "HTTP GET on behalf of an application, with the same address checks as the fetch_url tool (no private addresses, every redirect re-checked, size limit, per-host pace); recorded",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The response, whatever its HTTP status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Refused: address, scheme or header not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Fetch is not enabled in this Core",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The request failed (DNS, connection, timeout, too many redirects)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}